scripts/dev/NewbieLoraTrainer/requirements.txt
diffusers>=0.27.0
transformers>=4.38.0
peft>=0.8.2
accelerate>=0.27.0
safetensors>=0.4.0  
lycoris-lora>=3.4.0  
bitsandbytes>=0.42.0
toml>=0.10.2
Pillow>=10.2.0
huggingface-hub>=0.20.0
torchdiffeq>=0.2.0
timm